Created classes for each labeled individual for Noctua display
Though it feels awkward it actually does smell alright.. We are after
all really talking about classes of e.g. reactions not instances.  SO
having a representation of the class makes some sense.  Would be better
of they were classes from a structured ontology but this is a start.
And it makes the rendered models much more interesting to look at in
Noctua!