Added a default layout algorithm
sometimes it works pretty well.. other times gets overrun by giant
complexes.  

Switched from reaction/function part_of pathway/process to pathway
has_part function. With reasoning on it should be the same but is
convenient for accessing things from the pathway down.

When complex components are all in the same location as the complex,
delete them and leave the annotation on the complex.