fixes #21
This should result in stable URIs for biological processes and molecular
functions converted from Reactome BioPax.  These are the crucial
entities for cross-model linking coming up in #43 .  It is possible we
could increase the stability of entity URIs across different runs by
replacing all of the makeRandomIRI calls with something less random.  

The getEntityReferenceId method is currently totally reactome specific.
This will need adaptations to work on other BioPAX.